+++ tests/test_encode.py
@@ -400,7 +400,7 @@ def test_header_negative_incorrect_encoding():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .unknown encoding: invalid.: field h",
+        match="Failed to encode field, unknown encoding specified: field h",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -409,9 +409,6 @@ def test_header_negative_incorrect_ascii_data():
     """
     ASCII header is required by spec and provided.
     However, the data is not ASCII
-    CPython and PyPy throw differently worded exception
-    CPython: 'ascii' codec can't encode characters in position 0-5: ordinal not in range(128)
-    PyPy:    'ascii' codec can't encode character '\\xff' in position 0: ordinal not in range(128)
     """
     spec["h"]["data_enc"] = "ascii"
     spec["h"]["len_type"] = 0
@@ -426,7 +423,7 @@ def test_header_negative_incorrect_ascii_data():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .'ascii' codec can't encode character.*: ordinal not in range.128..: field h",
+        match="Failed to encode field, invalid data: field h",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -446,7 +443,7 @@ def test_header_negative_incorrect_bcd_data():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .non-hexadecimal number found in fromhex.. arg at position 0.: field h",
+        match="Failed to encode field, non-hex data: field h",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -646,9 +643,6 @@ def test_variable_header_bdc_over_max():
 def test_variable_header_bdc_odd():
     """
     BDC variable header is required and odd length is provided
-    CPython and PyPy throw differently worded exception
-    CPython: non-hexadecimal number found in fromhex() arg at position 5
-    PyPy:    non-hexadecimal number found in fromhex() arg at position 4
     """
     spec["h"]["data_enc"] = "b"
     spec["h"]["len_enc"] = "b"
@@ -661,7 +655,7 @@ def test_variable_header_bdc_odd():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .non-hexadecimal number found in fromhex.. arg at position 4|5.: field h",
+        match="Failed to encode field, odd-length hex data: field h",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -816,7 +810,7 @@ def test_variable_header_incorrect_encoding():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode length .unknown encoding: invalid.: field h",
+        match="Failed to encode field length, unknown encoding specified: field h",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -896,10 +890,7 @@ def test_type_ascii_over_max():
 def test_type_ascii_incorrect_data():
     """
     ASCII message type is required and provided.
-    However, the data is not ASCII
-    CPython and PyPy throw differently worded exception
-    CPython: 'ascii' codec can't encode characters in position 0-3: ordinal not in range(128)
-    PyPy:    'ascii' codec can't encode character '\\xff' in position 0: ordinal not in range(128)
+    However, the data is not ASCII.
     """
     spec["h"]["data_enc"] = "ascii"
     spec["h"]["len_type"] = 0
@@ -914,7 +905,7 @@ def test_type_ascii_incorrect_data():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .'ascii' codec can't encode character.*: ordinal not in range.128..: field t",
+        match="Failed to encode field, invalid data: field t",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -1094,9 +1085,6 @@ def test_type_bdc_over_max():
 def test_type_bdc_odd():
     """
     BDC message type is required and odd length is provided
-    CPython and PyPy throw differently worded exception
-    CPython: non-hexadecimal number found in fromhex() arg at position 3
-    PyPy:    non-hexadecimal number found in fromhex() arg at position 2
     """
     spec["h"]["data_enc"] = "ascii"
     spec["h"]["len_type"] = 0
@@ -1108,7 +1096,7 @@ def test_type_bdc_odd():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .non-hexadecimal number found in fromhex.. arg at position 2|3.: field t",
+        match="Failed to encode field, odd-length hex data: field t",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -1128,7 +1116,7 @@ def test_type_bdc_non_hex():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .non-hexadecimal number found in fromhex.. arg at position 3.: field t",
+        match="Failed to encode field, non-hex data: field t",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -1180,7 +1168,7 @@ def test_type_incorrect_encoding():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .unknown encoding: invalid.: field t",
+        match="Failed to encode field, unknown encoding specified: field t",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -1341,7 +1329,7 @@ def test_primary_bitmap_incorrect_encoding():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .unknown encoding: invalid.: field p",
+        match="Failed to encode field, unknown encoding specified: field p",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -1363,7 +1351,7 @@ def test_secondary_bitmap_incorrect_encoding():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .unknown encoding: invalid.: field 1",
+        match="ailed to encode field, unknown encoding specified: field 1",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -1630,9 +1618,6 @@ def test_fixed_field_ascii_incorrect_data():
     """
     ASCII fixed field is required and provided.
     However, the data is not ASCII
-    CPython and PyPy throw differently worded exception
-    CPython: 'ascii' codec can't encode characters in position 0-1: ordinal not in range(128)
-    PyPy:    'ascii' codec can't encode character '\\xff' in position 0: ordinal not in range(128)
     """
     spec["h"]["data_enc"] = "ascii"
     spec["h"]["len_type"] = 0
@@ -1651,7 +1636,7 @@ def test_fixed_field_ascii_incorrect_data():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .'ascii' codec can't encode character.*: ordinal not in range.128..: field 2",
+        match="Failed to encode field, invalid data: field 2",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -1943,9 +1928,6 @@ def test_fixed_field_bdc_over_max():
 def test_fixed_field_bdc_odd():
     """
     BDC fixed field is required and odd length is provided
-    CPython and PyPy throw differently worded exception
-    CPython: non-hexadecimal number found in fromhex() arg at position 5
-    PyPy:    non-hexadecimal number found in fromhex() arg at position 4
     """
     spec["h"]["data_enc"] = "ascii"
     spec["h"]["len_type"] = 0
@@ -1960,7 +1942,7 @@ def test_fixed_field_bdc_odd():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .non-hexadecimal number found in fromhex.. arg at position 4|5.: field 2",
+        match="Failed to encode field, odd-length hex data: field 2",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -1983,7 +1965,7 @@ def test_fixed_field_bdc_non_hex():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .non-hexadecimal number found in fromhex.. arg at position 2.: field 2",
+        match="Failed to encode field, non-hex data: field 2",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -2085,7 +2067,7 @@ def test_fixed_field_incorrect_encoding():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .unknown encoding: invalid.: field 2",
+        match="Failed to encode field, unknown encoding specified: field 2",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -2322,9 +2304,6 @@ def test_variable_field_bdc_over_max():
 def test_variable_field_bdc_odd():
     """
     BDC variable field is required and odd length is provided
-    CPython and PyPy throw differently worded exception
-    CPython: non-hexadecimal number found in fromhex() arg at position 5
-    PyPy:    non-hexadecimal number found in fromhex() arg at position 4
     """
     spec["h"]["data_enc"] = "ascii"
     spec["h"]["len_type"] = 0
@@ -2340,7 +2319,7 @@ def test_variable_field_bdc_odd():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode .non-hexadecimal number found in fromhex.. arg at position 4|5.: field 2",
+        match="Failed to encode field, odd-length hex data: field 2",
     ):
         iso8583.encode(doc_dec, spec=spec)
 
@@ -2526,6 +2505,6 @@ def test_variable_field_incorrect_encoding():
 
     with pytest.raises(
         iso8583.EncodeError,
-        match="Failed to encode length .unknown encoding: invalid.: field 2",
+        match="Failed to encode field length, unknown encoding specified: field 2",
     ):
         iso8583.encode(doc_dec, spec=spec)